# EaaS-openAPI.yaml
openapi: 3.0.0
info:
  title: Experimentation as a Service API Specification
  version: "1.0.0"
  description: |
    The Experimentation as a Service (EaaS) API Specification defines the APIs for onboarding applications and experiments, managing the experiment lifecycle, and the APIs exposed by the southbound plugin.
      - Application Package Management Interface:
         - Create Application Package Info
         - Upload Application Package
         - Fetch Application Package Info
         - Fetch Application Package Artifact
      - Experiment Descriptor Management Interface:
         - Create Experiment Descriptor Info
         - Upload Experiment Descriptor Archive
      - Experiment Lifecycle Management Interface:
         - Create Experiment Identifier
         - Instantiate Experiment
servers:
  - url: "{apiRoot}/eaas-application-repository/api/v1"
    variables:
      apiRoot:
        default: http://localhost:8080
        description: API root, defined by the service provider, e.g. `api.example.com` or `api.example.com/somepath`
  - url: "{apiRoot}/eaas-experimentation-descriptor-manager/v0.1"
    variables:
      apiRoot:
        default: http://localhost:9091
        description: API root, defined by the service provider, e.g. `api.example.com` or `api.example.com/somepath`
  - url: "{apiRoot}/eaas-experimentation-lifecycle-manager/v0.1"
    variables:
      apiRoot:
        default: http://localhost:9091
...
paths:
...
  /app_packages:
    get:
      tags: 
        - EaaS Application Repository
      summary: Query Application packages information.
      description: |
        Creates a new Application Package Info resource.
        According to ETSI NFV-SOL 005, the request payload is minimal and includes only userDefinedData.
      parameters:
      - in: query
        name: showPublic
        schema:
          type: boolean
        description: Returns also public applications (default false)
      responses:
        '200':
          description: Application Package Info created successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppPkgInfo'
        '4XX':
          description: 400 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '5XX':
          description: 500 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags: 
        - EaaS Application Repository
      summary: Create Application Package Info
      description: |
        Creates a new Application Package Info resource.
        According to ETSI NFV-SOL 005, the request payload is minimal and includes only userDefinedData.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppPkgInfoRequest'
      responses:
        '201':
          description: Application Package Info created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppPkgInfo'
        '4XX':
          description: 400 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '5XX':
          description: 500 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /app_packages/{appPkgId}:
    get:
      tags: 
        - EaaS Application Repository
      summary: Fetch Application Package Info
      description: Retrieves information about a specific Application package.
      parameters:
        - name: appPkgId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Application Package Info retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppPkgInfo'
        '4XX':
          description: 400 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '5XX':
          description: 500 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags: 
        - EaaS Application Repository
      summary: Delete Application Package Info
      description: Retrieves information about a specific Application package.
      parameters:
        - name: appPkgId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Application Package Info retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppPkgInfo'
        '4XX':
          description: 400 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '5XX':
          description: 500 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /app_packages/{appPkgId}/package_content:
    put:
      tags: 
        - EaaS Application Repository
      summary: Upload Application Package
      description: Uploads the Application package content (e.g. binary archive).
      parameters:
        - name: appPkgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/zip:
            schema:
              type: string
              format: binary
      responses:
        '202':
          description: Application package upload accepted.
        '4XX':
          description: 400 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '5XX':
          description: 500 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /app_packages/{appPkgId}/app_descriptor:
    get:
      tags: 
        - EaaS Application Repository
      summary: Read Application Descriptor of an on-boarded Application package.
      description: The GET method reads the content of the Application Descriptor within a Application package.
      parameters:
        - name: appPkgId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Shall be returned when the content of the Application Descriptor has been read successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppDescriptor'
        '4XX':
          description: 400 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '5XX':
          description: 500 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /app_packages/{appPkgId}/manifest:
    get:
      tags: 
        - EaaS Application Repository
      summary: Read the manifest of an on-boarded Application package
      description: The GET method reads the content of the manifest within a Application package.
      parameters:
        - name: appPkgId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Shall be returned when the content of the manifest has been read successfully.
          content:
            application/text:
              schema:
                type: string
                format: string
  /app_packages/{appPkgId}/artifacts/{artifactPath}:
    get:
      tags: 
        - EaaS Application Repository
      summary: Fetch Application Package Artifact
      description: Retrieves a specific artifact from a Application package.
      parameters:
        - name: appPkgId
          in: path
          required: true
          schema:
            type: string
        - name: artifactPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artifact fetched successfully.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '4XX':
          description: 400 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '5XX':
          description: 500 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /app_packages/{appPkgId}/ws:
    get:
      tags: 
        - EaaS Application Repository
      summary: Listen to app package update events
      description: Open a WebSocket to retrieve application update events (AppPkgEvent).
      parameters:
        - name: appPkgId
          in: path
          required: true
          schema:
            type: string
      responses:
        '101':
          description: Switching Protocol.
        '4XX':
            description: 400 family error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetails'
        '5XX':
          description: 500 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /app_packages/{appPkgId}/usage/{requestId}:
    put:
      tags: 
        - EaaS Application Repository
      summary: Request to use the AppPkgId
      description: The request ID will be associated with the app package ID, and the app package’s packageUsageState will be set to IN_USE
      parameters:
        - name: appPkgId
          in: path
          required: true
          schema:
            type: string
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content.
        '4XX':
          description: 400 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '5XX':
          description: 500 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags: 
        - EaaS Application Repository
      summary: Delete a previous usage request
      description: The request ID will no longer be associated with the app package ID. If there are no other requests on the package, the packageUsageState will be set to NOT_IN_USE.
      parameters:
        - name: appPkgId
          in: path
          required: true
          schema:
            type: string
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content.
        '4XX':
          description: 400 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '5XX':
          description: 500 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'





  # Experiment Descriptor Management Interface
  /experiment_descriptors:
    post:
      tags: 
        - EaaS Experimentation Description Manager
      summary: Create Experiment Descriptor Info
      description: |
        Creates a new Experiment Descriptor Info resource.
        The payload follows ETSI guidelines and uses a minimal model with userDefinedData.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExperimentDescriptorInfoRequest'
      responses:
        '201':
          description: Experiment Descriptor Info created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentDescriptorInfo'
  /experiment_descriptors/{experimentDescriptorId}:
    get:
      tags: 
        - EaaS Experimentation Description Manager
      summary: Read information about an individual Experiment descriptor resource.
      description: The GET method reads information about an individual Experiment descriptor.
      parameters:
        - name: experimentDescriptorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Shall be returned when information about the individual Experiment descriptor has been read successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentDescriptorInfo'
  /experiment_descriptors/{experimentDescriptorId}/experimentDescriptor_archive_content:
    put:
      tags: 
        - EaaS Experimentation Description Manager
      summary: Upload Experiment Descriptor Archive
      description: Uploads the Experiment Descriptor archive content.
      parameters:
        - name: experimentDescriptorId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/zip:
            schema:
              type: string
              format: binary
      responses:
        '202':
          description: Experiment Descriptor archive upload accepted.

  # Experiment Lifecycle Management Interface
  /experiment_instances:
    post:
      tags: 
        - EaaS Experimentation Lifecycle Manager
      summary: Create Experiment Identifier
      description: |
        The POST method creates a new Experiment instance resource.
      requestBody:
        description: The Experiment creation parameters.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExperimentRequest'
      responses:
        '201':
          description: >
            Shall be returned when a new "Individual Experiment instance" 
            resource and the associated Experiment instance identifier 
            has been created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentInstance'
  /experiment_instances/{experimentInstanceId}/instantiate:
    post:
      tags: 
        - EaaS Experimentation Lifecycle Manager
      summary: Instantiate Experiment
      description: Initiates the instantiation of an Experiment instance.
      parameters:
        - name: experimentInstanceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstantiateExperimentRequest'
      responses:
        '202':
          description: Experiment instantiation initiated.
  /experiment_instances/{experimentInstanceId}/terminate:
    post:
      tags: 
        - EaaS Experimentation Lifecycle Manager
      summary: Terminate Experiment
      description: Initiates the termination of an Experiment instance.
      parameters:
        - name: experimentInstanceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminateExperimentRequest'
      responses:
        '202':
          description: Experiment termination initiated.
  /application_onboarding:
    post:
      tags: 
        - Southbound Plugin
      summary: Onboard the application
      description: It will call the underlying Site APIs to onboard the application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppPkgInfo'
      responses:
        '201':
          description: Application onboarding accepted.
          content:
            application/text:
              schema:
                type: string
                description: Application onboarding id
  /create_application_instance:
    post:
      tags: 
        - Southbound Plugin
      summary: Create (and start) an application's instance
      description: It will call the underlying Site APIs to start the given instance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstanceApplicationRequest'
      responses:
        '201':
          description: Experiment's instance operation starting accepted.
          content:
            application/text:
              schema:
                type: string
                description: Application instance id
  /stop_application_instance:
    post:
      tags: 
        - Southbound Plugin
      summary: Stop an application's instance
      description: It will call the underlying Site APIs to stop the given instance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StopInstanceApplicationRequest'
      responses:
        '202':
          description: Experiment's instance operation stopping accepted.
  /{instanceId}/state/ws:
    get:
      tags: 
        - Southbound Plugin
      summary: Listen to instance state events
      description: Open a WebSocket to retrieve the instance state (string).
      parameters:
        - name: instanceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '101':
          description: Switching Protocol.
        '4XX':
            description: 400 family error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetails'
        '5XX':
          description: 500 family error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  schemas:
...
    StopInstanceApplicationRequest:
      type: object
      description: Data Structure for request a stop of a Application Instance
      properties:
        appInstanceId:
          type: string
          description: Identifier of the instance of the Application. This identifier is allocated by the Southbound Plugin.
      required:
        - appInstanceId
    ProblemDetails:
      type: object
      description: Data structure from IETF RFC 7807
      properties:
        type:
          type: string
          description: >
            A URI reference according to IETF RFC 3986 [3] that identifies
            the problem type. It is encouraged that the URI provides
            human-readable documentation for the problem (e.g. using
            HTML) when dereferenced. When this member is not present,
            its value is assumed to be "about:blank".
        title:
          type: string
          description: >
            A short, human-readable summary of the problem type. It should
            not change from occurrence to occurrence of the problem,
            except for purposes of localization. If type is given and other
            than "about:blank", this attribute shall also be provided.
        status:
          type: integer
          description: The HTTP status code for this occurrence of the problem.
          example: 403
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem. It may yield further information if dereferenced.
        additionalAttribute:
          type: object
          description: Any number of additional attributes, as defined in a specification or by an implementation.
          additionalProperties: true
      required:
        - status
        - detail
...
    CreateAppPkgInfoRequest:
      type: object
      description: Request payload for creating Application Package Info (minimal ETSI model).
      properties:
        userDefinedData:
          type: object
          description: User defined data for the Application package.
          additionalProperties: true
...
    CreateExperimentDescriptorInfoRequest:
      type: object
      description: Request payload for creating Experiment Descriptor Info (minimal ETSI model).
      properties:
        userDefinedData:
          type: object
          description: Arbitrary user-defined data.
          additionalProperties: true
      required:
        - userDefinedData
...
    CreateExperimentRequest:
      type: object
      description: The Experiment creation parameters.
      properties:
        experimentDescriptorId:
          type: string
          description: Identifier of the Experiment Descriptor that defines the Experiment instance to be created.
        experimentName:
          type: string
          description: Human-readable name of the Experiment instance to be created.
        experimentDescription:
          type: string
          description: Human-readable description of the Experiment instance to be created.
      required:
        - experimentDescriptorId
        - experimentName
        - experimentDescription
...
    AppDescriptor:
      type: object
      description: Representation of a Application Descriptor Info resource.
      properties:
        appDescriptorId:
          type: string
          description: >
            Identifier of this Application descriptor information element. This attribute shall be globally unique. The format will be
            defined in the data model specification phase.
        appDescriptorExtInvariantId:
          type: string
          description: >
            Identifies a Application Descriptor in a version independent manner. This attribute is invariant across versions of the Application Descriptor
            that fulfil certain conditions related to the external connectivity and management of the Application.
        appProvider:
          type: string
          description: Provider of the Application and of the Application Descriptor
        appProductName:
          type: string
          description: Name to identify the Application Product. Invariant for the Application Product lifetime.
        appSoftwareVersion:
          type: string
          description: >
            Software version of the Application. This is changed when there is any change to the software that is included in the Application Package.
        appDescriptorVersion:
          type: string
          description: Specifies the version of the Application Descriptor
        appmInfo:
          type: array
          description: Specifies Application M(s) compatible with the Application described in this version of the Application Descriptor.
          items:
            type: string
        swImageDesc:
          type: array
          description: Defines descriptors of software images to be used by the Application.
          items:
            $ref: '#/components/schemas/SwImageDesc'
        vdu:
          type: array
          description: VDUs related to this Application Descriptor.
          items:
            $ref: '#/components/schemas/VDU'
        deploymentFlavour:
          type: array
          description: Describes specific DF(s) of a Application with specific requirements for capacity and performance.
          items:
            $ref: '#/components/schemas/DeploymentFlavour'
        appExtCpd:
          type: array
          description: Describes external interface(s) exposed by this Application enabling connection with a VL.
          items:
            $ref: '#/components/schemas/AppExtCpd'
        virtualCpd:
          type: array
          description: >
            Describes a virtual connection point allowing to access a set of Application Component instances (based on their respective VDUs)
          items:
            $ref: '#/components/schemas/VirtualCpd'
      required:
        - appDescriptorId
        - appDescriptorExtInvariantId
        - appProvider
        - appProductName
        - appSoftwareVersion
        - appDescriptorVersion
    
### Experiment model
...